Moving render structure into a separate file
- As a refactor, the Renderer object has been moved into a separate
  file to make it easier to maintain.
