+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod figure;
+pub mod renderer;
 mod utils;
